requirement.txt
adblockparser==0.7
alembic==1.4.0
appdirs==1.4.3
beautifulsoup4==4.9.1
blinker==1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
comtypes==1.1.7
cryptography==2.8
decorator==4.4.2
distlib==0.3.0
extraction==0.3
feedparser==6.0.1
filelock==3.0.12
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Login==0.4.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
flask-whooshalchemy3==0.2.0
google==3.0.0
gTTS==2.1.1
gTTS-token==1.1.3
html5lib==1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.3
jsonify==0.5
lxml==4.5.2
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
networkx==2.5
nexmo==2.4.0
nltk==3.4.5
numpy==1.19.2
passlib==1.7.2
pbr==3.1.1
pdf2image==1.14.0
Pillow==7.2.0
pycparser==2.20
PyJWT==1.7.1
pypiwin32==223
pytesseract==0.3.6
python-dateutil==2.8.1
python-editor==1.0.4
pyttsx3==2.90
pytz==2019.3
pywin32==228
quora==0.1.23
requests==2.23.0
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.0.1
speech==0.5.2
SQLAlchemy==1.3.12
stripe==2.44.0
summary-extraction==0.2
textblob==0.15.3
urllib3==1.25.8
virtualenv==20.0.16
w3lib==1.22.0
webencodings==0.5.1
Werkzeug==0.16.0
Whoosh==2.7.4
wikipedia==1.4.0
WTForms==2.2.1
